Move executor package into trac-lib-exec
