+++ trac-libs/trac-lib-exec/src/main/java/com/accenture/trac/common/exec/kubernetes/KubernetesBatchState.java
@@ -14,16 +14,9 @@
  * limitations under the License.
  */
 
-package com.accenture.trac.svc.orch.exec;
+package com.accenture.trac.common.exec.kubernetes;
 
-import com.accenture.trac.api.JobStatusCode;
-import com.accenture.trac.config.JobResult;
+import com.accenture.trac.common.exec.ExecutorState;
 
-public class ExecutorPollResult {
-
-    public String jobKey;
-    public JobStatusCode statusCode;
-    public String errorMessage;
-    public JobResult jobResult;
-    public ExecutorState executorState;
+public class KubernetesBatchState extends ExecutorState {
 }
